refactor code because we need to follow the correct index topics
